# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: Stellar-IndigoPay API
  version: 1.0.0
  description: |
    OpenAPI 3.0 documentation for the backend REST API used by Stellar-IndigoPay.
    The API exposes project discovery, donations, profiles, leaderboard, notifications,
    ratings, updates, jobs, stats, and admin management endpoints.
  contact:
    name: Stellar-IndigoPay Team
    url: https://github.com/Stellar-IndigoPay/Stellar-IndigoPay
  license:
    name: MIT
servers:
  - url: http://localhost:4000
    description: Local development server  
  - url: https://api.stellarindigopay.com
    description: Production server
...
security:
  - BearerAuth: []
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    RefreshCookie:
      type: apiKey
      in: cookie
      name: refresh_token
      description: Rotating admin refresh token. httpOnly, SameSite=Strict, scoped to /api.
...
  responses:
    BadRequest:
      description: The request payload or query parameters are invalid.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Unauthorized:
      description: Authentication is required or the provided token is invalid.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
    NotFound:
      description: The requested resource was not found.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    RateLimitExceeded:
      description: The request has been rate-limited. All endpoints carry rate-limit response headers (X-RateLimit-Limit, X-RateLimit-Remaining, X-RateLimit-Reset) and return 429 with a Retry-After header when the limit is exceeded. See the API documentation for per-endpoint tiers.
      headers:
        Retry-After:
          description: Number of seconds to wait before retrying.
          schema:
            type: integer
        X-RateLimit-Limit:
          description: Maximum requests allowed (or bucket capacity for token-bucket endpoints).
          schema:
            type: integer
        X-RateLimit-Remaining:
          description: Requests remaining in the current window (or tokens left in the bucket).
          schema:
            type: integer
        X-RateLimit-Reset:
          description: Epoch seconds when the limit resets (or next token becomes available).
          schema:
            type: integer
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: string
                example: Too many requests — Try again later.
              retryAfter:
                type: integer
                description: Epoch seconds after which a retry may succeed.
                example: 1742169602
...
  schemas:
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
          example: Invalid request payload
      required:
        - error
...
    AdminLoginResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: object
          properties:
            token:
              type: string
              description: Access token. Hold it in memory only.
            expiresIn:
              type: integer
              description: Access token lifetime in seconds.
    AdminRefreshResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: object
          properties:
            token:
              type: string
            expiresIn:
              type: integer
    AdminSessionsResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: array
          items:
            type: object
            properties:
              id:
                type: string
                format: uuid
                description: Refresh token family identifying the session.
              createdAt:
                type: string
                format: date-time
              expiresAt:
                type: string
                format: date-time
              current:
                type: boolean
                description: True for the session making this request.
    AdminMeResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: object
          properties:
            username:
              type: string
            role:
              type: string
    AdminAuditLogResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: array
          items:
            type: object
            properties:
              id:
                type: string
              actor:
                type: string
              action:
                type: string
              target_type:
                type: string
              target_id:
                type: string
              metadata:
                type: object
              ip_address:
                type: string
              created_at:
                type: string
                format: date-time
        total:
          type: integer
        page:
          type: integer
        pageSize:
          type: integer
...
    AdminLoginRequest:
      type: object
      required: [username, password]
      properties:
        username:
          type: string
        password:
          type: string
...
paths:
...
  /api/admin/login:
    post:
      tags: [Admin]
      summary: Log in as an administrator
      operationId: adminLogin
      description: >-
        Authenticate as an administrator using username and password
        credentials. Returns a 15-minute access token and sets the rotating
        refresh token as an httpOnly, SameSite=Strict cookie scoped to /api.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminLoginRequest"
      responses:
        "200":
          description: Admin access token. The refresh token is set as a cookie, never returned in the body.
          headers:
            Set-Cookie:
              description: refresh_token=<opaque>; HttpOnly; SameSite=Strict; Path=/api; Max-Age=604800
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminLoginResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "429":
          $ref: "#/components/responses/RateLimitExceeded"
  /api/admin/refresh:
    post:
      tags: [Admin]
      summary: Rotate the admin session
      operationId: adminRefreshToken
      description: >-
        Exchange the refresh_token cookie for a new access token. Each call
        revokes the presented refresh token and issues a replacement in the
        same family. Replaying an already-revoked token revokes every session
        in that family.
      security:
        - RefreshCookie: []
      responses:
        "200":
          description: Rotated session.
          headers:
            Set-Cookie:
              description: The replacement refresh token.
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminRefreshResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "429":
          $ref: "#/components/responses/RateLimitExceeded"
  /api/admin/logout:
    post:
      tags: [Admin]
      summary: End the admin session
      operationId: adminLogout
      description: >-
        Revoke the refresh token family behind the refresh_token cookie,
        blacklist the presented access token until its natural expiry, and
        clear the cookie. Always succeeds.
      security:
        - BearerAuth: []
        - RefreshCookie: []
      responses:
        "200":
          description: Session ended.
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
        "429":
          $ref: "#/components/responses/RateLimitExceeded"
  /api/admin/sessions:
    get:
      tags: [Admin]
      summary: List active admin sessions
      operationId: listAdminSessions
      description: List the authenticated admin's active sessions, one per refresh token family.
      responses:
        "200":
          description: Active sessions.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminSessionsResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /api/admin/sessions/{id}/revoke:
    post:
      tags: [Admin]
      summary: Revoke an admin session
      operationId: revokeAdminSession
      description: Revoke every refresh token in the given session family.
      parameters:
        - name: id
          in: path
          required: true
          description: Session id (refresh token family).
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Session revoked.
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "429":
          $ref: "#/components/responses/RateLimitExceeded"
  /api/admin/me:
    get:
      tags: [Admin]
      summary: Get current admin identity
      operationId: getAdminMe
      description: Retrieve the currently authenticated admin user's identity and role.
      responses:
        "200":
          description: Current admin identity.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminMeResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
...
  /api/admin/audit-log:
    get:
      tags: [Admin]
      summary: Query admin audit log
      operationId: queryAuditLog
      description: Query the admin audit log with optional filters for actor, action, and pagination.
      parameters:
        - name: actor
          in: query
          schema:
            type: string
        - name: action
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: pageSize
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 200
            default: 50
      responses:
        "200":
          description: Audit log page.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminAuditLogResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
...